refactor: extract password hashing function into separate class and inject into services
