+++ medallert-backend/src/api.ts
@@ -3,7 +3,7 @@ import { hello } from "./routes/hello.js";
 import { configureOpenAPIDocs } from "./routes/open-api.js";
 import { authMiddleware } from "./routes/middleware/auth-middleware.js";
 import { auth } from "./routes/auth/index.js";
-import { passwordRecovery } from "./routes/password-recovery.js";
+import { passwordRecovery } from "./routes/password-recovery/index.js";
 
 export const app = new Hono();
