+++ medallert-backend/src/routes/password-recovery/index.ts
@@ -1,0 +1,8 @@
+import { Hono } from "hono";
+import { changePassword } from "./change-password.js";
+import { requestRecovery } from "./recover-password.js";
+
+export const passwordRecovery = new Hono();
+
+passwordRecovery.route("/", requestRecovery);
+passwordRecovery.route("/", changePassword);
